create calendar for visitor AND refacto js calendar
