+++ app/controllers/api/v1/coaches_controller.rb
@@ -1,5 +1,5 @@
 class Api::V1::CoachesController < Api::V1::BaseController
   def index
-    @coaches = policy_scope(Coach)
+    @coaches = policy_scope(User, policy_scope_class: CoachPolicy::Scope)
   end
 end
