+++ app/policies/coach_policy.rb
@@ -1,7 +1,7 @@
 class CoachPolicy < ApplicationPolicy
   class Scope < Scope
     def resolve
-      scope.all
+      scope.coach
     end
   end
 end
