+++ app/views/bookings/index.html.erb
@@ -17,8 +17,8 @@
 <% end %>
 <div class="container px-5">
   <%= link_to "Créer nouvelle disponibilité", new_booking_path, class: 'btn btn-success py-3 text-white d-flex justify-content-center'  %>
-  <div id="calendar" class="m-3">
 
+  <div id="calendar-coach" class="m-3">
   </div>
 
 </div>
