+++ app/views/coaches/show.html.erb
@@ -40,6 +40,10 @@
           <a class="btn btn-lg btn-primary" href="#">Vérifier la disponilité</a>
         </div>
       </div>
+      <div class="card" style="width: 18rem;">
+        <div id="calendar">
+        </div>
+      </div>
     </div>
   </div>
 </div>
